refactor: move DEFAULT_APPROVAL_EXPIRY_SECONDS to config.constants.investigation
opensre.py was removed on main; relocate the constant to investigation.py
(which already owns tool-framework-adjacent limits) and update all callers.
