--- config/constants/opensre.py
@@ -1,10 +1,0 @@
-"""OpenSRE product version constants."""
-
-from __future__ import annotations
-
-from typing import Final
-
-DEFAULT_RELEASE_VERSION: Final[str] = "0.1"
-
-# Approval tokens auto-expire after this many seconds (5 minutes).
-DEFAULT_APPROVAL_EXPIRY_SECONDS: Final[int] = 300
